rest_yab/requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-filter==22.1
mysqlclient==2.1.1
uritemplate==4.1.1